move shared crate-details functions into database crate
