+++ src/db/delete.rs
@@ -1,11 +1,10 @@
 use crate::{Config, error::Result};
 use anyhow::Context as _;
+use docs_rs_database::crate_details::update_latest_version_id;
 use docs_rs_storage::{AsyncStorage, rustdoc_archive_path, source_archive_path};
 use docs_rs_types::{CrateId, Version};
 use fn_error_context::context;
 use sqlx::Connection;
-
-use super::update_latest_version_id;
 
 /// List of directories in docs.rs's underlying storage (either the database or S3) containing a
 /// subdirectory named after the crate. Those subdirectories will be deleted.
