+++ src/db/mod.rs
@@ -1,5 +1,4 @@
 //! Database operations
-pub use self::add_package::update_latest_version_id;
 pub(crate) use self::add_package::{
     add_doc_coverage, finish_build, finish_release, initialize_build, initialize_crate,
     initialize_release, update_build_with_error,
