+++ src/web/crate_details.rs
@@ -20,6 +20,7 @@ use axum::{
 };
 use chrono::{DateTime, Utc};
 use docs_rs_cargo_metadata::{Dependency, ReleaseDependencyList};
+use docs_rs_database::crate_details::{Release, latest_release, parse_doc_targets};
 use docs_rs_headers::CanonicalUrl;
 use docs_rs_registry_api::OwnerKind;
 use docs_rs_storage::{AsyncStorage, PathNotFoundError};
@@ -75,33 +76,6 @@ struct RepositoryMetadata {
     forks: i32,
     issues: i32,
     name: Option<String>,
-}
-
-#[derive(Debug, Clone, Eq, PartialEq)]
-pub(crate) struct Release {
-    pub id: ReleaseId,
-    pub version: Version,
-    #[allow(clippy::doc_overindented_list_items)]
-    /// Aggregated build status of the release.
-    /// * no builds -> build In progress
-    /// * any build is successful -> Success
-    ///   -> even with failed or in-progress builds we have docs to show
-    /// * any build is failed -> Failure
-    ///   -> we can only have Failure or InProgress here, so the Failure is the
-    ///      important part on this aggregation level.
-    /// * the rest is all builds are in-progress -> InProgress
-    ///   -> if we have any builds, and the previous conditions don't match, we end
-    ///      up here, but we still check.
-    ///
-    /// calculated in a database view : `release_build_status`
-    pub build_status: BuildStatus,
-    pub yanked: Option<bool>,
-    pub is_library: Option<bool>,
-    pub rustdoc_status: Option<bool>,
-    pub target_name: Option<String>,
-    pub default_target: Option<String>,
-    pub doc_targets: Option<Vec<String>>,
-    pub release_time: Option<DateTime<Utc>>,
 }
 
 impl CrateDetails {
@@ -211,7 +185,7 @@ impl CrateDetails {
             rustdoc_status: krate.rustdoc_status,
             target_name: krate.target_name.clone(),
             default_target: krate.default_target,
-            doc_targets: krate.doc_targets.map(MetaData::parse_doc_targets),
+            doc_targets: krate.doc_targets.map(parse_doc_targets),
             yanked: krate.yanked,
             rustdoc_css_file: krate
                 .rustc_version
@@ -368,65 +342,6 @@ impl CrateDetails {
     pub fn latest_release(&self) -> Result<&Release> {
         latest_release(&self.releases).ok_or_else(|| anyhow!("crate without releases"))
     }
-}
-
-pub(crate) fn latest_release(releases: &[Release]) -> Option<&Release> {
-    if let Some(release) = releases.iter().find(|release| {
-        release.version.pre.is_empty()
-            && release.yanked == Some(false)
-            && release.build_status != BuildStatus::InProgress
-    }) {
-        Some(release)
-    } else {
-        releases
-            .iter()
-            .find(|release| release.build_status != BuildStatus::InProgress)
-    }
-}
-
-/// Return all releases for a crate, sorted in descending order by semver
-pub(crate) async fn releases_for_crate(
-    conn: &mut sqlx::PgConnection,
-    crate_id: CrateId,
-) -> Result<Vec<Release>, anyhow::Error> {
-    let mut releases: Vec<Release> = sqlx::query!(
-        r#"SELECT
-             releases.id as "id: ReleaseId",
-             releases.version as "version: Version",
-             release_build_status.build_status as "build_status!: BuildStatus",
-             releases.yanked,
-             releases.is_library,
-             releases.rustdoc_status,
-             releases.release_time,
-             releases.target_name,
-             releases.default_target,
-             releases.doc_targets
-         FROM releases
-         INNER JOIN release_build_status ON releases.id = release_build_status.rid
-         WHERE
-             releases.crate_id = $1"#,
-        crate_id.0,
-    )
-    .fetch(&mut *conn)
-    .try_filter_map(|row| async move {
-        Ok(Some(Release {
-            id: row.id,
-            version: row.version,
-            build_status: row.build_status,
-            yanked: row.yanked,
-            is_library: row.is_library,
-            rustdoc_status: row.rustdoc_status,
-            target_name: row.target_name,
-            default_target: row.default_target,
-            doc_targets: row.doc_targets.map(MetaData::parse_doc_targets),
-            release_time: row.release_time,
-        }))
-    })
-    .try_collect()
-    .await?;
-
-    releases.sort_by(|a, b| b.version.cmp(&a.version));
-    Ok(releases)
 }
 
 #[derive(Debug, Clone, Template)]
@@ -727,7 +642,7 @@ mod tests {
         fake_release_that_failed_before_build,
     };
     use anyhow::Error;
-    use docs_rs_database::testing::TestDatabase;
+    use docs_rs_database::{crate_details::releases_for_crate, testing::TestDatabase};
     use docs_rs_registry_api::CrateOwner;
     use docs_rs_types::KrateName;
     use http::StatusCode;
